Add test for bag-management

+++ backend/libs/pickme/logic/Bags.js
@@ -1,5 +1,6 @@
 const Database = require('../database/Database');
 const NotificationScheduler = require('./NotificationScheduler');
+const BagFactory = require('./BagFactory');
 
 /**
  * Manages the bags.
@@ -11,7 +12,7 @@ class Bags {
   static async getBags() {
     const dbClient = await Database.connect();
     try {
-      return await dbClient.getBags();
+      return (await dbClient.getBags()).map((bag) => BagFactory.create(bag));
     } finally {
       await dbClient.close();
     }
@@ -19,17 +20,26 @@ class Bags {
 
   /**
    * Updates an existing bag.
-   * @param {Bag} bag An existing bag (must contain an id).
-   * @return {Promise<Bag>} The new version of the bag.
+   * @param {Bag} rawBag An existing bag (must contain an id).
+   * @return {Promise<{error: Error, value: Bag}>}
+   * The new version of the bag.
    */
-  static async updateBag(bag) {
+  static async updateBag(rawBag) {
+    const bag = BagFactory.create(rawBag);
+    if (!bag) {
+      return {
+        error: new Error(`${rawBag} is malformed Bag`),
+        value: undefined,
+      };
+    }
+
     const dbClient = await Database.connect();
     dbClient.startTransaction();
     try {
       await dbClient.updateBag(bag);
       NotificationScheduler.scheduleNotification(bag);
       await dbClient.commitTransaction();
-      return bag;
+      return {value: bag};
     } finally {
       await dbClient.close();
     }
@@ -37,17 +47,26 @@ class Bags {
 
   /**
    * Adds a new bag.
-   * @param {Bag} bag An existing bag (must not contain an id).
-   * @return {Promise<Bag>} The newly created bag (contains the assigned id).
+   * @param {Bag} rawBag An existing bag (must not contain an id).
+   * @return {Promise<{error: Error, value: Bag}>}
+   * The newly created bag (contains the assigned id).
    */
-  static async addBag(bag) {
+  static async addBag(rawBag) {
+    const bag = BagFactory.create(rawBag);
+    if (!bag) {
+      return {
+        error: new Error(`${rawBag} is malformed Bag`),
+        value: undefined,
+      };
+    }
+
     const dbClient = await Database.connect();
     dbClient.startTransaction();
     try {
       const newBag = await dbClient.addBag(bag);
       NotificationScheduler.scheduleNotification(bag);
       await dbClient.commitTransaction();
-      return newBag;
+      return {value: newBag};
     } finally {
       await dbClient.close();
     }
@@ -56,17 +75,28 @@ class Bags {
   /**
    * Deletes an existing bag.
    * @param {string} bagId The id of an existing bag.
-   * @return {Promise<boolean>}
+   * @return {Promise<{error: Error, value: boolean}>}
    * True, if a bag has been deleted, otherwise false.
    */
   static async deleteBag(bagId) {
+    if (!await Database.isObjectIdValid(bagId)) {
+      return {
+        error: new Error(`${bagId} is invalid ObjectId`),
+        value: false,
+      };
+    }
+
     const dbClient = await Database.connect();
     dbClient.startTransaction();
     try {
       const itemDeleted = await dbClient.deleteBag(bagId);
-      NotificationScheduler.abortScheduledNotification(bagId);
+      if (itemDeleted) {
+        NotificationScheduler.abortScheduledNotification(bagId);
+      }
       await dbClient.commitTransaction();
-      return itemDeleted;
+      return {
+        value: itemDeleted,
+      };
     } finally {
       await dbClient.close();
     }
@@ -75,12 +105,26 @@ class Bags {
   /**
    * Get bag by id.
    * @param {string} bagId The id of the bag.
-   * @return {Promise<Bag>} The Bag.
+   * @return {Promise<{error: Error, value: Bag}>}
+   * Bag or undefined, if there is no bag with the specified id.
    */
   static async getBag(bagId) {
+    if (!await Database.isObjectIdValid(bagId)) {
+      return {
+        error: new Error(`${bagId} is invalid ObjectId`),
+        value: undefined,
+      };
+    }
+
     const dbClient = await Database.connect();
     try {
-      return dbClient.getBag(bagId);
+      const rawBag = await dbClient.getBag(bagId);
+      if (rawBag) {
+        const bag = await BagFactory.create(rawBag);
+        return {value: bag};
+      } else {
+        return {value: undefined};
+      }
     } finally {
       await dbClient.close();
     }
